feat: (widget) Add style with light,dark themes, and previousPrayer
Improved the design, and also fetched the previous prayer, this design
tries to be similar to the one on iOS.

I'm not sure how we can set the theme, I'm assuming it will follow the
device, but would be good to be able to configure it.

Also, the time shown in the widget is not correct, I think we need to
play around timezones.

Then, next we need to have a dynamic counter